+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added examples for time picker on `DatePickerInput`
+
 ### Fixed
 
 ## [9.4.0] 2023-04-11